% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{env_clear}
\alias{env_clear}
\title{Clear Environment Data}
\usage{
env_clear(envir = .GlobalEnv)
}
\arguments{
\item{envir}{Environment, which environment you wish to clear}
}
\value{
Invisible NULL
}
\description{
Clear Environment Data
}
