% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{env_restore}
\alias{env_restore}
\title{Restore Saved Environment}
\usage{
env_restore(
  from = ".envirs/",
  file = get_latest_file(from),
  rm_all = TRUE,
  envir = .GlobalEnv
)
}
\arguments{
\item{from}{Character, directory where environment data is stored. Defaults
to \verb{.envirs/}}

\item{file}{Character, which file in \code{from} contains the environment you wish
to restore. Defaults to the most recently modified}

\item{rm_all}{Logical, whether or not to remove current environment data
before loading data from new environment. Defaults to \code{TRUE}}

\item{envir}{Environment, where data will be loaded. Defaults to \code{.GlobalEnv}}
}
\value{
Loads all variables to \code{envir}
}
\description{
Restore Saved Environment
}
