% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{env_save}
\alias{env_save}
\title{Snapshot of Global Environment}
\usage{
env_save(to = ".envirs/", envir = .GlobalEnv)
}
\arguments{
\item{to}{Character, file path where the environment data will be saved. Defaults to \verb{.envirs/}}

\item{envir}{Environment, which environment you want to save. Defaults to \code{.GlobalEnv}}
}
\value{
Invisible, location of the file containing saved environment data
}
\description{
Snapshot of Global Environment
}
