% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{get_latest_file}
\alias{get_latest_file}
\title{Get Name of The Most Recently Saved File}
\usage{
get_latest_file(from = ".envirs/")
}
\arguments{
\item{from}{Character, directory in which files are searched}
}
\value{
Name of file most recently saved
}
\description{
Get Name of The Most Recently Saved File
}
