% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Anti-Value Matching}
\usage{
x \%notin\% y
}
\arguments{
\item{x}{Vector or NULL: the values to be matched. Long vectors are supported.}

\item{y}{Vector or NULL: the values to be matched against. Long vectors are not supported.}
}
\value{
Logical, whether or not \code{x} was in \code{y}. See \code{\%in\%} documentation for more information.
}
\description{
Match returns a vector of the positions of (first) anti-matches of its first argument in its second.
}
