% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%remove\%}
\alias{\%remove\%}
\title{Remove Patterns Within String}
\usage{
x \%remove\% pattern
}
\arguments{
\item{x}{Character vector.}

\item{pattern}{Character string containing a regular expression to be matched and removed from \code{x}.}
}
\value{
See \code{\%in\%} documentation for more information.
}
\description{
\code{gsub} with \code{ignore.case = TRUE, perl = TRUE, fixed = FALSE}.
}
