% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%//\%}
\alias{\%//\%}
\title{Concatenate Strings}
\usage{
x \%//\% y
}
\arguments{
\item{x}{Character, first part of a string to be concatenated to the
beginning of \code{y}}

\item{y}{Character, second part f a string to be concatenated to the end of
\code{x}}
}
\value{
Character, concatenated strings \code{x} and \code{y}, with "/" added or
removed when necessary
}
\description{
Concatenate filepaths
}
