% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see.R
\name{see}
\alias{see}
\title{Wrapper for \code{View} - Invoke Data Viewer}
\usage{
see(x)
}
\arguments{
\item{x}{an R object which can be coerced to a data frame with non-zero numbers of rows and columns.}
}
\value{
Invisible NULL. The functions puts up a window and returns immediately: the window can be closed via its controls or menus.
}
\description{
see(x) = View(x). See \code{View} documentation with \code{?View} for more information.
}
