% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{env_save}
\alias{env_save}
\title{Snapshot of Global Environment}
\usage{
env_save(environment = .GlobalEnv)
}
\arguments{
\item{environment}{The environment to be saved}
}
\value{
A copy of the global environment
}
\description{
Snapshot of Global Environment
}
